requirements.txt
pycurl
pyopenssl
pycryptodome